Update NavigationMenu namespace, moving into Models
