+++ packages/framework/tests/Feature/NavigationMenuTest.php
@@ -3,14 +3,14 @@
 namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Framework\Hyde;
+use Hyde\Framework\Models\NavigationMenu;
 use Hyde\Framework\Models\NavItem;
-use Hyde\Framework\Modules\Navigation\NavigationMenu;
 use Hyde\Framework\Modules\Routing\Route;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
 
 /**
- * @covers \Hyde\Framework\Modules\Navigation\NavigationMenu
+ * @covers \Hyde\Framework\Models\NavigationMenu
  */
 class NavigationMenuTest extends TestCase
 {
